custom-elements/custom-element-lifecycle/types-of-callbacks/attribute-changed-callback-test-003.html

<!DOCTYPE html>
<html>
<head>
<title>Test attribute removing to check attributeChanged callback of a custom element</title>
<meta name="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<meta name="assert" content="attributeChanged callback must be enqueued whenever custom element's attribute is added, changed or removed">
<link rel="help" href="http://www.w3.org/TR/custom-elements/#types-of-callbacks">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src='../../testcommon.js'></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script type="text/javascript">

test(function() {
    var doc = newHTMLDocument();
    var callbackCalledCounter = 0;

    var proto = Object.create(HTMLElement.prototype);
    proto.attributeChangedCallback  = function () {
        callbackCalledCounter++;
    }
    var GeneratedConstructor = doc.registerElement('x-a-1', {prototype: proto});
    var customElement = new GeneratedConstructor();
    //attributeChangedCallback should be called first time here
    customElement.setAttribute('class', 'someClass');
  //attributeChangedCallback should be called second time here
    customElement.removeAttribute('class');

    assert_equals(callbackCalledCounter, 2, 'Callback attributeChanged was called ' +
            'wrong number of times');
}, 'Test attributeChanged callback of the custom element created in JavaScript. ' +
    'Test that callback is called when attribute is removed');


test(function() {
    var doc = newHTMLDocument();

    HTML5_ELEMENTS.forEach(function(element) {
        var callbackCalledCounter = 0;
        var obj = doc.createElement(element);

        var proto = Object.create(obj.constructor.prototype);
        proto.attributeChangedCallback  = function () {
            callbackCalledCounter++;
        }
        var GeneratedConstructor = doc.registerElement('x-' + element + '-' + element + '-1', {
            prototype: proto,
            extends: element
        });
        var customElement = new GeneratedConstructor();
        //attributeChangedCallback should be called first time here
        customElement.setAttribute('class', 'someClass');
      //attributeChangedCallback should be called second time here
        customElement.removeAttribute('class');

        assert_equals(callbackCalledCounter, 2, 'Callback attributeChanged was called ' +
                'wrong number of times for "' + element + '"');
    });
}, 'Test attributeChanged callback of the custom element that extends HTML element. ' +
    'Test that callback is called when attribute is removed');


test(function() {
    var doc = newHTMLDocument();
    var callbackCalledCounter = 0;

    var proto = Object.create(HTMLElement.prototype);
    proto.attributeChangedCallback = function () {
        callbackCalledCounter++;
    }
    doc.registerElement('x-a-2', {prototype: proto});
    doc.body.innerHTML = '<x-a-2 id="xa2" class="theClass"></x-a-2>';
    var customElement = doc.querySelector('#xa2');
    customElement.removeAttribute('class');
    assert_equals(callbackCalledCounter, 1, 'Callback attributeChanged was called ' +
            'wrong number of times');
}, 'Test attributeChanged callback of the custom element created in HTML. ' +
    'Test that callback is called when attribute is removed');


test(function() {
    var doc = newHTMLDocument();
    var argument1 = 'arg1';
    var argument2 = 'arg2';
    var argument3 = 'arg3';

    var proto = Object.create(HTMLElement.prototype);
    proto.attributeChangedCallback = function(arg1, arg2, arg3) {
        argument1 = arg1;
        argument2 = arg2;
        argument3 = arg3;
    }
    doc.registerElement('x-a-3', {prototype: proto});
    doc.body.innerHTML = '<x-a-3 id="xa3" class="theClass"></x-a-3>';
    var customElement = doc.querySelector('#xa3');
    customElement.removeAttribute('class');

    assert_equals(argument1, 'class', 'Wrong value of the first argument of ' +
        'attributeChanged callback');
    assert_equals(argument2, 'theClass', 'Wrong value of the second argument of ' +
        'attributeChanged callback');
    assert_equals(argument3, null, 'Wrong value of the third argument of ' +
        'attributeChanged callback');
}, 'Test attributeChanged callback of the custom element created in JavaScript. ' +
    'Test callback arguments when attribute is removed');


test(function() {
    var doc = newHTMLDocument();
    var callbackCalledCounter = 0;

    var proto = Object.create(HTMLElement.prototype);
    proto.attributeChangedCallback = function () {
        callbackCalledCounter++;
    }
    doc.body.innerHTML = '<x-a-4 id="xa4" class="theClass"></x-a-4>';
    var customElement = doc.querySelector('#xa4');
    // this should not call or enquee attributeChangedCallback
    customElement.setAttribute('class', 'someClass');
    // this one too
    customElement.removeAttribute('class');

    doc.registerElement('x-a-4', {prototype: proto});
    // this call invokes attributeChangedCallback
    customElement.setAttribute('name', 'someName');
    // and this one
    customElement.removeAttribute('name');

    assert_equals(callbackCalledCounter, 2, 'Callback attributeChanged was called ' +
            'wrong number of times');
}, 'Test attributeChanged callback of the custom element created in HTML. ' +
    'Test attribute removing when element in unresolved at first');


testInIFrame('../../resources/x-element.html', function(context) {
    var doc = context.iframes[0].contentDocument;
    var callbackCalledCounter = 0;

    var proto = Object.create(HTMLElement.prototype);
    proto.attributeChangedCallback = function () {
        callbackCalledCounter++;
    }
    doc.registerElement('x-element', {prototype: proto});
    var customElement = doc.querySelector('#xelement');
    customElement.setAttribute('class', 'someClass');
    customElement.removeAttribute('class');

    assert_equals(callbackCalledCounter, 2, 'Callback attributeChanged was called ' +
            'wrong number of times');
}, 'Test attributeChanged callback for the document that has browsing context. ' +
    'Test that callback is called wheeb attribute is removed');


testInIFrame('../../resources/x-element.html', function(context) {
    var doc = context.iframes[0].contentDocument;
    var argument1 = 'arg1';
    var argument2 = 'arg2';
    var argument3 = 'arg3';

    var proto = Object.create(HTMLElement.prototype);
    proto.attributeChangedCallback = function(arg1, arg2, arg3) {
        argument1 = arg1;
        argument2 = arg2;
        argument3 = arg3;
    }
    doc.registerElement('x-element', {prototype: proto});
    doc.body.innerHTML = '<x-element id="xelement" class="theClass"></x-element>';
    var customElement = doc.querySelector('#xelement');
    customElement.removeAttribute('class');

    assert_equals(argument1, 'class', 'Wrong value of the first argument of ' +
        'attributeChanged callback');
    assert_equals(argument2, 'theClass', 'Wrong value of the second argument of ' +
        'attributeChanged callback');
    assert_equals(argument3, null, 'Wrong value of the third argument of ' +
        'attributeChanged callback');
}, 'Test attributeChanged callback for the document that has browsing context. ' +
    'Test that callback arguments');


testInIFrame('../../resources/x-element.html', function(context) {
    var doc = context.iframes[0].contentDocument;
    var callbackCalledCounter = 0;

    var proto = Object.create(HTMLElement.prototype);
    proto.attributeChangedCallback = function () {
        callbackCalledCounter++;
    }
    var customElement = doc.querySelector('#xelement');
    // this should not call or enquee attributeChangedCallback
    customElement.setAttribute('name', 'someName');
    // this one too
    customElement.removeAttribute('name');

    doc.registerElement('x-element', {prototype: proto});
    // this call invokes attributeChangedCallback
    customElement.setAttribute('class', 'someClass');
    // this call invokes attributeChangedCallback at second time
    customElement.removeAttribute('class');

    assert_equals(callbackCalledCounter, 2, 'Callback attributeChanged was called ' +
            'wrong number of times');
}, 'Test attributeChanged callback for the document that has browsing context. ' +
    'Test attribute removing when element in unresolved at first');

</script>
</body>
</html>
